Add new attribute, used_by, to namelist options needed by MPAS dycore in CAM
This commit adds the attribute used_by="cam" to namelist options need by
the MPAS-A dycore in CAM. A Python script (build_cam_namelists.py) included
with the MPAS-A dycore in CAM may be used to extract all such namelist options
and to build pieces of files needed by CAM to define namelists.

For the CAM-MPAS Python script to extract an option, the enclosing nml_record
tag must have used_by="cam", as must the nml tag itself.