Remove unused arguments scalar_tend and rho_zz_int from atm_advance_scalars
Following a comment in the code, the unused arguments scalar_tend and rho_zz_int
can be removed from the argument list of atm_advance_scalars. Furthermore, as
the actual argument (scalar_tend_array) to scalar_tend was only used in the call
to atm_advance_scalars, it can also be removed from the atm_time_integration
module.