Define role constants in (almost) a single location
The objective of this commit is to avoid having definitions of role constants
(e.g., ROLE_INTEGRATE and ROLE_RADIATION) in multiple locations, all of which
need to be kept consistent.

Now, role constants are defined in core-specific code guarded by preprocessing
directives in the mpas_intracoupler module. The exception is the use of role
constants in mpas_pool_routines, where a dependence on mpas_intracoupler would
create a circular dependence (mpas_dmpar -> mpas_pool_routines ->
mpas_intracoupler -> mpas_dmpar).