Add role_is and role_includes functions to mpas_coupler
The mpas_coupler type now includes two new functions, role_is and role_includes,
that are set to point to core-specific functions component_role_is and
component_role_includes. The new functions may be used to make code more
readable by avoiding bit-testing operations in core code.

For example, if roles are implemented as bit-masks, the following would
determine if a task's role included ROLE_FOO:

    if (iand(mpas_cpl % role(), ROLE_FOO) == ROLE_FOO) then

Instead, the above code can now be replaced with:

    if (mpas_cpl % role_includes(ROLE_FOO)) then

An alternative to

    if (iand(mpas_cpl % role() == ROLE_FOO) then

that is now available is

    if (iand(mpas_cpl % role_is(ROLE_FOO)) then

.