Uncomment call to summarize_timestep in MPAS-A dycore
The summarize_timestep routine may still not give correct printout when running
on GPUs, since there still may be fields that need to be copied from device to
host memory. However, for CPU-only runs, the timestep summary can be helpful for
debugging or checking on the progress of a running simulation.