Use default initialization to give attType in att_list_type an invalid value
In order to ensure that attributes that have not been assigned valid contents
are not inadvertently used, this commit gives the attType member of
the att_list_type a default initialization of -1, which does not match any
valid attribute type.

An example of where lack of initialization can lead to errors is in
the mpas_io_streams module, where an attribute list is written to an output
stream.

Beginning around line 4323 in mpas_io_streams.F, a 'select case' construct
is executed for all nodes, including the head node, of a field's attribute list:

      att_cursor => attList
      if (ioDirection == MPAS_IO_WRITE) then
         do while (associated(att_cursor))
            select case (att_cursor % attType)
               case (MPAS_ATT_INT)
                  call MPAS_io_put_att(handle, trim(att_cursor % attName), att_cursor % attValueInt, fieldname)
               case (MPAS_ATT_INTA)
                  call MPAS_io_put_att(handle, trim(att_cursor % attName), att_cursor % attValueIntA, fieldname)
               case (MPAS_ATT_REAL)
                  call MPAS_io_put_att(handle, trim(att_cursor % attName), att_cursor % attValueReal, fieldname)
               case (MPAS_ATT_REALA)
                  call MPAS_io_put_att(handle, trim(att_cursor % attName), att_cursor % attValueRealA, fieldname)
               case (MPAS_ATT_TEXT)
                  call MPAS_io_put_att(handle, trim(att_cursor % attName), att_cursor % attValueText, fieldname)
            end select
            att_cursor => att_cursor % next
         end do

Typically, fields are allocated an attribute list head node that does not
contain a valid attribute. If the head node of an attribute list happened to
have an attType that matches MPAS_ATT_INT, MPAS_ATT_INTA, etc., then
the stream would attempt to write that invalid attribute node to a file.