Explain the utility of mpas_constants_compute_derived in its comment block
Since it may not otherwise be obvious why the mpas_constants_compute_derived
routine was added, this commit adds an explanation of the utility of this
routine in its documentation comments.

This commit also reverts the change in the previous commit to the copyright date
in the mpas_constants module, since updating copyright dates is probably
something that should be coordinated across source files.