Revert "update NF_ to NF90_ in mpas_io.F"
This reverts commit 291176d4d689de195e101e553b0d077b360f17b2.

The change from "NF_FILL_*" to "NF90_FILL_*" constants in mpas_io.F was needed
when building MPAS-Atmosphere as a CAM dycore in the past, but since CAM/CESM
has updated to using PIO2, these changes are no longer needed (because the
affected code is pre-processed out when USE_PIO2 is defined).

Since the use of "NF_FILL_*" constants worked without apparent problem in stand-
alone MPAS, perhaps as a consequence of the way that PIO was installed by MPAS
users, it seems preferable to revert the mpas_io.F code to its former state.