Convert tend_{ru,rtheta,rho}_physics to scratch arrays and allocate as needed
In order to allow stand-alone MPAS-Atmosphere to allocate the tend_ru_physics,
tend_rtheta_physics, and tend_rho_physics fields only for the duration of a call
to atm_srk3 (as was done when these fields were allocatable arrays), this commit
converts these fields to scratch fields. The allocation and deallocation of
these fields in the atm_srk3 routine only happens when MPAS_CAM_DYCORE is not
defined.

When MPAS-Atmosphere is run as a dycore in CAM, the tend_{ru,rtheta,rho}_physics
fields need to persist between calls to the dycore, since these fields are set
in the CAM physics-dynamics coupling layer. When MPAS_CAM_DYCORE is defined,
these scratch fields are allocated and deallocated once per run in the
mpas_atm_dynamics_init and mpas_atm_dynamics_finalize routines.

Note that CAM-MPAS will need to ensure that calls to mpas_atm_dynamics_init and
mpas_atm_dynamics_finalize are made at appropriate points in the CAM-MPAS run
sequence.