Add acc exit data statements to deallocate_scratch routines
Scratch variables may be fetched and used in OpenACC parallel sections,
but can cause partially present issues when the array is deallocated but
the memory isn't freed on the device. By using an `!$acc exit data
delete(f % array) finalize` statement in the deallocate_scratch
routines, the memory is guranateed to be freed on the device too.